refactor: 상영회 등록하기 로직 훅으로 분리
