+++ src/screens/screening/writing/WritingScreen.tsx
@@ -1,6 +1,5 @@
-import {useEffect, useRef, useState} from 'react';
+import {useEffect, useRef} from 'react';
 import {TextInput, View} from 'react-native';
-import {DateParsable} from 'react-native-calendar-picker';
 
 import DefaultContainer from '@/components/container/defaultContainer';
 import Typography from '@/components/typography';
@@ -13,18 +12,14 @@ import TextArea from '@/components/inputs/textArea';
 import DismissKeyboardView from '@/components/dismissKeyboardView';
 import Input from '@/components/input';
 import BoxButton from '@/components/buttons/boxButton';
-
+import useHandleInput from './hooks/useHandleInput';
 import useScreeningMutation from '@/hooks/mutaions/useScreeningMutation';
 import {ScreenRouteProp} from '@/types/navigator';
 import useNavigator from '@/hooks/useNavigator';
 import stackScreens from '@/constants/stackScreens';
 import {KorCategoryValues} from '@/models/enums/category';
-import {IScreeningBodyRequest} from '@/models/screening/request/screeningRequestDto';
 
 import {writingStyles} from './WritingScreen.style';
-import {useQuery} from '@tanstack/react-query';
-import {getScreeningMyDetailContent} from '@/apis/screening/detail';
-import {getCategory} from '@/utils/getCategory';
 
 interface IWritingScreenProps {
   route: ScreenRouteProp<'WritingScreen'>;
@@ -33,58 +28,12 @@ const WritingScreen = ({route: {params}}: IWritingScreenProps) => {
   const {type, search, id} = params;
   const {uploadScreening, modifyScreening} = useScreeningMutation();
   const {stackNavigation} = useNavigator();
-  const [inputValues, setInputValues] = useState<IScreeningBodyRequest>({
-    posterImgUrl: '',
-    screeningTitle: '',
-    hostName: '',
-    category: '',
-    screeningStartDate: undefined,
-    screeningEndDate: undefined,
-    screeningStartTime: undefined,
-    location: search,
-    information: '',
-    formUrl: '',
-    hostPhoneNumber: '',
-    hostEmail: '',
-    hasAgreed: false,
-  });
-
-  const {data} = useQuery({
-    queryKey: ['screeningMyDetail'],
-    queryFn: () => {
-      if (id) {
-        return getScreeningMyDetailContent(id);
-      }
-    },
-    enabled: type === 'post',
-  });
+  const {setModify, inputValues, setInputValues, onChangeInput} =
+    useHandleInput(search, id);
 
   useEffect(() => {
     onChangeInput('location', search);
   }, [search]);
-
-  useEffect(() => {
-    if (data && id && type === 'modified') {
-      console.log(data.data);
-      setInputValues({
-        ...inputValues,
-        posterImgUrl: data.data.posterImgUrl,
-        screeningTitle: data.data.screeningTitle,
-        hostName: data.data.hostName,
-        category: getCategory(data.data.category),
-        screeningStartDate: new Date(data.data.screeningStartDate),
-        screeningEndDate: new Date(data.data.screeningEndDate),
-        screeningStartTime: new Date(data.data.screeningStartTime),
-        information: data.data.information,
-        formUrl: data.data.formUrl,
-        hostPhoneNumber: data.data.hostPhoneNumber
-          ? data.data.hostPhoneNumber
-          : '',
-        hostEmail: data.data.hostEmail,
-        hasAgreed: data.data.hasAgreed,
-      });
-    }
-  }, [setInputValues]);
 
   const {
     screeningTitle,
@@ -98,13 +47,6 @@ const WritingScreen = ({route: {params}}: IWritingScreenProps) => {
     hasAgreed,
     posterImgUrl,
   } = inputValues;
-
-  const onChangeInput = (
-    inputName: string,
-    value: string | DateParsable | undefined | boolean | Date,
-  ) => {
-    setInputValues({...inputValues, [inputName]: value});
-  };
 
   const canGoNext =
     posterImgUrl &&
@@ -133,10 +75,10 @@ const WritingScreen = ({route: {params}}: IWritingScreenProps) => {
     if (type === 'post') {
       await uploadScreening.mutateAsync(inputValues);
     }
-    if (type === 'modified' && data) {
-      const body = {screeningId: data.data.screeningId, ...inputValues};
-      await modifyScreening.mutateAsync(body);
-    }
+    //if (type === 'modified' && data) {
+    //  const body = {screeningId: data.data.screeningId, ...inputValues};
+    //  await modifyScreening.mutateAsync(body);
+    //}
   };
 
   return (
